refactor: LottoCheckResponse 분리, equals,hashcode,toString 정의
